backend/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "biomarker-explorer-backend"
version = "0.1.0"
description = "Digital Biomarker Explorer — FastAPI backend for BHS sensor data"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.29",
  "pydantic>=2.6",
  "google-cloud-bigquery>=3.19",
  "google-cloud-storage>=2.14",
  "db-dtypes>=1.2",
  "pandas>=2.1",
  "numpy>=1.26",
  "scipy>=1.12",
  "pyarrow>=15",
  "python-dateutil>=2.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "pytest-asyncio>=0.23",
  "httpx>=0.27",
  "ruff>=0.3",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py310"
